2006-12-13 v1.6.1

+++ src/cps1/timer.c
@@ -55,7 +55,7 @@ static CPUINFO cpu[MAX_CPU];
 	ƒ[ƒJƒ‹•Ï”
 ******************************************************************************/
 
-static const float time_slice = (int)(1000000.0 / FPS);
+static const float time_slice = 1000000.0 / FPS;
 
 static float global_offset;
 static float base_time;
